revert to SPA FIXED
