--- client/src/views/createdocstore/index.js
@@ -1,2 +1,0 @@
-import CreateDocStore from './CreateDocStore';
-export default CreateDocStore;
