--- client/src/views/homepage/index.js
@@ -1,2 +1,0 @@
-import HomePage from './HomePage';
-export default HomePage;
